op-program: Don't close deriver when derivation goes idle
The FCU event will close the deriver in time. If we cannot reach the
target otherwise, it doesn't matter that we don't close.

In Holocene, if the last attributes are invalid, deposit-only attributes
are retried. Due to event race conditions, the deriver may then already
get idle while the engine is still processing events. So the program can
close without having updated the safe head to pending/local yet.