downgrade `SystemConfigInterop` back to proper version from `origin/develop`
The `SystemConfigInterop` contract was downgraded before
`op-contracts/v1.8.0-rc.1` to "fix CI." However, this improperly bumped
the contract's version for the release, and is causing the semver diff
checker to fail.

This change also corrects the `check-semver-diff.sh` script to account
for the change in directory structure in `packages/contracts-bedrock` in
`origin/develop` at the time of writing.